requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.7
